refactor(llmz): split runtime orchestration
